BkuR1QbW actualizacion de archivo de becas genera un nuevo archivo en vez de modificar el viejo. Problemas relacionados a concurrencia estan siendo tenidos en cuenta (se aplica operacion atomica a documento de mongodb findOneAndUpdate)

+++ server_node/apimaster/controllers/upload-handlers/update-scholarship-handler.js
@@ -4,16 +4,18 @@ const axios = require("axios");
 const studentFileMetadata = require("../../../models/student-file-metadata");
 const { PYFLASK_URL } = require("../../../config/config");
 const { NotFound } = require("../../../errors/errors-index");
+const { v4: uuidv4 } = require("uuid");
+const { PICKLE_SUFFIX } = require("../../../config/paths");
 
 const updateScholarshipsHandler = async (req, res) => {
-  let toBeUpdated = null;
+  let currentDataFile = null;
 
   try {
-    toBeUpdated = await studentFileMetadata.findOne({
+    currentDataFile = await studentFileMetadata.findOne({
       year: req.body.year,
       type: req.body.type,
     });
-    if (!toBeUpdated) {
+    if (!currentDataFile) {
       throw new NotFound(
         `No scholarship file found with year ${req.body.year} and type ${req.body.type}.`
       );
@@ -23,25 +25,76 @@ const updateScholarshipsHandler = async (req, res) => {
     throw error;
   }
 
+  const destinationFilename =
+    currentDataFile.year +
+    "_" +
+    currentDataFile.type +
+    "_" +
+    uuidv4() +
+    PICKLE_SUFFIX;
+
+  const destinationFilepath =
+    currentDataFile.folder + "/" + destinationFilename; //The complete file path for a new file that will contain the result of the operation
+
   try {
     await axios.post(PYFLASK_URL + "/conversions/studentscholarships/update", {
-      sourceFile: req.body.tempFolder + "/" + req.body.tempFilename,
-      destinationFile: toBeUpdated.folder + "/" + toBeUpdated.filename,
-      type: toBeUpdated.type,
+      newDataFile: req.body.tempFolder + "/" + req.body.tempFilename, //The file that holds the new data
+      currentDataFile: currentDataFile.folder + "/" + currentDataFile.filename, //The file that holds the current data
+      destinationFile:
+        //The file that will hold the combination of the new and old data
+        destinationFilepath,
+      type: currentDataFile.type,
     });
   } catch (error) {
     await fs.unlink(req.body.tempFolder + "/" + req.body.tempFilename);
     throw error;
   }
 
-  await fs.unlink(req.body.tempFolder + "/" + req.body.tempFilename);
-  let result = await studentFileMetadata.findOneAndUpdate(
-    { _id: toBeUpdated._id },
-    {},
-    { fields: "-folder -filename", new: true }
-  ); //Trigger a dummy update to update the updatedAt field with the current date.
+  let result;
+  try {
+    result = await studentFileMetadata.findOneAndUpdate(
+      //result will hold the old file data
+      { _id: currentDataFile._id },
+      { filename: destinationFilename }
+    );
+  } catch (error) {
+    //In case the db update fails, delete de temp file and the new file
+    try {
+      await fs.unlink(req.body.tempFolder + "/" + req.body.tempFilename);
+    } catch (error) {}
+    try {
+      await fs.unlink(destinationFilepath);
+    } catch (error) {}
+    throw error;
+  }
 
-  res.status(StatusCodes.OK).json({ success: true, result: result });
+  if (!result) {
+    //In case the file was not found , then it was deleted
+    try {
+      await fs.unlink(req.body.tempFolder + "/" + req.body.tempFilename);
+    } catch (error) {}
+    try {
+      await fs.unlink(destinationFilepath);
+    } catch (error) {}
+    throw new NotFound("The requested scholarship file was not found");
+  }
+
+  // If everything goes right, delete temp file and the old file
+  try {
+    await fs.unlink(req.body.tempFolder + "/" + req.body.tempFilename);
+  } catch (error) {}
+  try {
+    await fs.unlink(result.folder + "/" + result.filename);
+  } catch (error) {}
+
+  let old_version = result.toJSON(); //Do not send the folder and filename in response. Only JSON can be modified
+
+  delete old_version.filename;
+  delete old_version.folder;
+  res.status(StatusCodes.OK).json({
+    success: true,
+    old_version: old_version,
+  });
 };
 
 module.exports = updateScholarshipsHandler;
